requirement.txt
fastapi
uvicorn
python-multipart
python-dotenv
psycopg[binary]
PyMuPDF
langchain-text-splitters
sentence-transformers
camelot-py[cv]
opencv-python-headless
pandas
requests
httpx
groq
nest-asyncio
duckduckgo-search
torch
joblib
scikit-learn
numpy